requirements.txt
# requirements.txt
# --------------------
# This file records the packages and requirements needed in order for
# the library to work as expected. And to run tests.
aiohttp
fsspec
GDAL
frictionless
numpy
platformdirs
pygeoprocessing>=2.4.3
pyyaml
requests
